refactor : [ISSUE-4주차코드리뷰] Service - ServiceImpl 구조 제거 및 swagger 추가
