+++ build.gradle
@@ -24,6 +24,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.mapstruct:mapstruct:1.5.3.Final'
     implementation 'org.springframework.boot:spring-boot-starter-mail'
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
+    implementation 'commons-lang:commons-lang:2.6'
 
     // 쿼리 파라미터 값 확인 외부 라이브러리
     //implementation 'com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.5.7'
