+++ src/main/java/com/zerobase/task/invite/infra/mail/MailService.java
@@ -10,8 +10,6 @@ import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.mail.javamail.MimeMessagePreparator;
 import org.springframework.stereotype.Service;
 
-import java.nio.charset.StandardCharsets;
-
 @Slf4j
 @RequiredArgsConstructor
 @Service
