+++ src/main/resources/application.yml
@@ -3,6 +3,7 @@ server:
     encoding:
       charset: UTF-8
       force: true
+  port: 8080
 
 spring:
   h2:
@@ -21,7 +22,9 @@ spring:
   properties:
      hibernate:
        format_sql:true
-
+  mvc:
+    pathmatch:
+      matching-strategy: ant_path_matcher
   mail:
     host: smtp.gmail.com
     port: 587
